requirements.txt
Django>=6.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
psycopg2-binary>=2.9.11
django-filter>=25.2
pytest>=9.0.2
pytest-django>=4.11.1
factory-boy>=3.3.3
black>=25.12.0
flake8>=7.3.0
python-dotenv>=1.2.1
gunicorn>=23.0.0
whitenoise>=6.11.0
django-widget-tweaks>=1.5.0
drf-spectacular>=0.29.0
drf-spectacular-sidecar>=2025.12.1